requirements.txt
streamlit==1.32.0
django==5.0.3
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
yfinance==0.2.37
python-dotenv==1.0.1
scikit-learn==1.4.1.post1
finnhub-python==2.4.19
alpha_vantage==2.3.1
google-generativeai==0.3.2
groq==0.4.2
python-multipart==0.0.9
openpyxl==3.1.2
pdfplumber==0.10.3
requests==2.31.0
pydantic==2.6.3
pydantic-core==2.16.3